+++ app/src/pages/Create.jsx
@@ -1,0 +1,5 @@
+const Create = () => {
+  return <button>hi</button>;
+};
+
+export default Create;
